states object view

+++ api/v1/views/states.py
@@ -4,40 +4,69 @@ Module for handling all RESTFul
 API actions
 '''
 
-from flask import Flask, jsonify, abort
+from flask import Flask, jsonify, abort, request
 from models import storage
+from models.state import State
 import os
+from api.v1.views import app_views
+import json
 
 app = Flask(__name__)
 
-@app.route('/api/v1/states', methods=['GET'])
-def get_states():
-    """Retrieves the list of all State objects"""
-    states = storage.all('State')
-    return jsonify([state.to_dict() for state in states])
 
-@app.route('/api/v1/states/<state_id>', methods=['GET'])
-def get_state(state_id):
-    """Retrieves a state object"""
-    state = storage.get('State', state_id)
+@app_views.route('/states', methods=['GET', 'POST'], strict_slashes=False)
+def states():
+    """Handles GET (all states) and POST
+    (create a new state) requests."""
+    if request.method == 'GET':
+        states = storage.all(State).values()
+        return jsonify([state.to_dict() for state in states])
+
+    if request.method == 'POST':
+        if not request.get_json():
+            abort(400, description="Not a JSON")
+
+        data = request.get_json()
+
+        if 'name' not in data:
+            abort(400, description="Missing name")
+
+        new_state = State(**data)
+        storage.new(new_state)
+        storage.save()
+
+        return jsonify(new_state.to_dict()), 201
+
+
+@app_views.route('/states/<state_id>',
+                 methods=['GET', 'PUT', 'DELETE'], strict_slashes=False)
+def state(state_id):
+    """Handles GET, PUT (update),
+    and DELETE (delete) requests
+    for a specific state."""
+    state = storage.get(State, state_id)
+
     if state is None:
         abort(404)
-    return jsonify(state.to_dict())
 
-@app.route('/api/v1/states/<state_id>', methods=['DELETE'])
-def delete_state(state_id):
-    """Delete a state object"""
-    state = storage.get('State', state_id)
-    if state is None:
-        abort(404)
-    storage.delete(state)
-    storage.save()
-    return jsonify({}), 200
+    if request.method == 'GET':
+        return jsonify(state.to_dict())
 
-if __name__ == "__main__":
-    '''
-    Starting server
-    '''
-    host = os.getenv("HBNB_API_HOST", "0.0.0.0")
-    port = int(os.getenv("HBNB_API_PORT", 5000))
-    app.run(host=host, port=port, threaded=True)
+    if request.method == 'PUT':
+        if not request.get_json():
+            abort(400, description="Not a JSON")
+
+        data = request.get_json()
+
+        ignore_keys = ['id', 'created_at', 'updated_at']
+        for key, value in data.items():
+            if key not in ignore_keys:
+                setattr(state, key, value)
+
+        storage.save()
+        return jsonify(state.to_dict()), 200
+
+    if request.method == 'DELETE':
+        storage.delete(state)
+        storage.save()
+        return jsonify({}), 200
